add jdbcuserdao, and some tests

+++ src/test/resources/db/h2/schema.sql
@@ -1,5 +1,5 @@
 create table users (
-id int not null primary key auto_increment,
+id bigint not null primary key auto_increment,
 email varchar(32) not null unique,
 password varchar(32) not null,
 phone_number varchar(32) unique,
